requirements.txt
moviepy
gtts
pillow
numpy
openai
python-dotenv